refactor: move "@computed" decorator back to the @corpuscule/element as a separate module (#8)
