--- packages/decorator-computed/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@corpuscule/decorator-computed",
-  "version": "0.1.0",
-  "description": "Computing decorator for Corpuscule",
-  "main": "lib/index.js",
-  "module": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "author": "Vlad Rindevich <rindevich.vs@gmail.com> (https://github.com/Lodin)",
-  "bugs": {
-    "url": "https://github.com/corpusculejs/corpuscule/issues"
-  },
-  "homepage": "https://github.com/corpusculejs/corpuscule/packages/decorator-computed#readme",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/corpusculejs/corpuscule.git"
-  },
-  "dependencies": {
-    "@corpuscule/utils": "^0.3.0"
-  }
-}
