+++ scripts/project.js
@@ -1,7 +1,9 @@
 const packages = {
   context: ["index"],
-  "decorator-computed": ["index"],
-  element: ["index"],
+  element: [
+    "computed",
+    "index",
+  ],
   redux: ["index"],
   router: ["index"],
   styles: ["index"],
