+++ test/test.js
@@ -1,5 +1,4 @@
 import "../packages/context/__tests__";
-import "../packages/decorator-computed/__tests__";
 import "../packages/element/__tests__";
 import "../packages/redux/__tests__";
 import "../packages/router/__tests__";
